apollo_storage: small fix in add migration class to append block state diff

+++ crates/apollo_storage/src/state/mod.rs
@@ -503,7 +503,10 @@ impl StateStorageWriter for StorageTxn<'_, RW> {
         write_nonces(&thin_state_diff.nonces, &self.txn, block_number, &nonces_table)?;
 
         for (class_hash, _) in &thin_state_diff.declared_classes {
-            declared_classes_block_table.insert(&self.txn, class_hash, &block_number)?;
+            let not_declared = declared_classes_block_table.get(&self.txn, class_hash)?.is_none();
+            if not_declared {
+                declared_classes_block_table.insert(&self.txn, class_hash, &block_number)?;
+            }
         }
 
         write_compiled_class_hashes(
@@ -596,6 +599,7 @@ impl StateStorageWriter for StorageTxn<'_, RW> {
             &self.txn,
             &thin_state_diff,
             &declared_classes_block_table,
+            block_number,
         )?;
         let deleted_classes = delete_declared_classes(
             &self.txn,
@@ -781,11 +785,23 @@ fn delete_declared_classes_block<'env>(
     txn: &'env DbTransaction<'env, RW>,
     thin_state_diff: &ThinStateDiff,
     declared_classes_block_table: &'env DeclaredClassesBlockTable<'env>,
+    block_number: BlockNumber,
 ) -> StorageResult<Vec<ClassHash>> {
     let mut deleted_data = Vec::new();
     for class_hash in thin_state_diff.declared_classes.keys() {
-        declared_classes_block_table.delete(txn, class_hash)?;
-        deleted_data.push(*class_hash);
+        let class_block_entry =
+            declared_classes_block_table.get(txn, class_hash)?.ok_or_else(|| {
+                StorageError::DBInconsistency {
+                    msg: format!(
+                        "Attempting to revert declaration of class {class_hash} but it doesn't \
+                         exist in the DB"
+                    ),
+                }
+            })?;
+        if class_block_entry == block_number {
+            declared_classes_block_table.delete(txn, class_hash)?;
+            deleted_data.push(*class_hash);
+        }
     }
     Ok(deleted_data)
 }
